pyproject.toml
[project]
name = "hutch-bunny"
version = "1.0.0"
description = "A Cohort Discovery Task API worker"
license = { text = "MIT License" }
readme = "README.md"
authors = [
  { name = "Jon Couldridge", email = "jonathan.couldridge@nottingham.ac.uk" },
  { name = "Daniel Lea", email = "daniel.lea@nottingham.ac.uk" },
  { name = "Tri Thien Nguyen", email = "thien.nguyen1@nottingham.ac.uk" },
  { name = "James Mitchell-White", email = "james.mitchell-white1@nottingham.ac.uk" },
  { name = "Anwar Gaungoo", email = "anwar.gaungoo@nottingham.ac.uk" },
  { name = "Vasiliki Panagi", email = "vasiliki.panagi@nottingham.ac.uk" },
]
maintainers = [
  { name = "Jon Couldridge", email = "jonathan.couldridge@nottingham.ac.uk" },
]
requires-python = ">=3.13"
dependencies = [
  "numpy>=2.2.1",
  "pandas>=2.2.3",
  "psycopg[binary]>=3.2.3",
  "python-dotenv>=1.0.1",
  "requests>=2.32.3", # daemon-only
  "sqlalchemy>=2.0.36",
  "trino>=0.331.0", # should become optional
  "pymssql>=2.3.2", # for SQL Server
  "pydantic-settings>=2.7.1",
  "tenacity>=9.1.2", # for retry
]

[project.scripts]
bunny = "hutch_bunny.cli:main"
bunny-daemon = "hutch_bunny.daemon:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.8.6",
    "pytest>=8.3.4",
    "pandas-stubs>=2.2.3.241126",
    "mypy>=1.14.1",
    "types-requests>=2.32.0.20241016",
    "coverage>=7.6.12",
    "pytest-cov>=6.0.0",
    "pre-commit>=4.1.0",
    "types-python-dateutil>=2.9.0.20250516",
]

[tool.pytest.ini_options]
markers = [
    "unit: Unit tests",
    "integration: Integration tests (DB, API, etc.)",
    "end_to_end: End-to-end tests (CLI, Daemon)",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.13"
strict = true
plugins = ['pydantic.mypy']

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true

# Prevent ANY implicit `Any` types  
disallow_any_generics = true
disallow_any_explicit = true
disallow_any_unimported = true
disallow_any_expr = false
disallow_any_decorated = true

# Ensure everything is typed properly  
disallow_untyped_defs = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_globals = true
disallow_incomplete_defs = true
check_untyped_defs = true

# Enforce subclassing & generics rules  
disallow_subclassing_any = true

# Prevent implicit behaviors  
implicit_reexport = false
strict_equality = true
no_implicit_optional = true

# Maximize warnings & errors  
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true
warn_unreachable = true
warn_redundant_casts = true

# Ensure explicit enum values  
strict_concatenate = true

# Miscellaneous  
show_error_codes = true
pretty = true
